+++ spec/amakanize/author_name_spec.rb
@@ -12,6 +12,7 @@ RSpec.describe Amakanize::AuthorName do
       "ぽんかん(8)" => "ぽんかん8",
       "ぽんかん8" => "ぽんかん8",
       "ぽんかん⑧" => "ぽんかん8",
+      "漫画:ハノカゲ" => "ハノカゲ",
     }.each do |key, value|
       context "with #{key.inspect}" do
         let(:raw) do