+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.2
 - Remove parentheses of numericals in author name
+- Remove role name from author name
 
 ## 0.0.1
 - 1st release :tada: